+++ sys/arch/x86_64/x86_64.h
@@ -110,6 +110,7 @@ u64 x86_64_ring0_stack_top(void);
 void x86_64_map_user_page(u64 va, u64 phys, int writable);
 void x86_64_map_user_page_to(u64 pml4_phys, u64 va, u64 phys, int writable);
 u64 x86_64_create_user_space(void);
+void x86_64_free_user_space(u64 pml4_phys); /* reclaim a user AS (execve teardown) */
 void x86_64_set_user_kstack(u64 top);
 void x86_64_syscall_init(void); /* enable SYSCALL/SYSRET (STAR/LSTAR/SFMASK/EFER.SCE) */
 void x86_64_syscall(struct x86_64_trapframe *tf);